+++ jobmon_gui/src/types/TaskInstance.ts
@@ -11,6 +11,9 @@ export type TaskInstance = {
     ti_stdout: string | null;
     ti_stdout_log: string | null;
     ti_wallclock: string | number | null;
+    ti_submit_date: string | null;
+    ti_status_date: string | null;
+    ti_queue_name: string | null;
 };
 
 export type TypeInstanceResponse = {